Added prompt and interface
I redesigned how the prompts are created and fixed up some of the code. Everything is a lot cleaner with an interface class now acting as an intermidiate event handler for everything. The prompt is missing options like moving where the cursor is located. This may never be able to be implemented unless I modify the interface.