Uses the default cursor
I relized that the system broke from the setup outputing the last line with a line return. This offset the whole outputer which I fixed at one point. This had made the previous cursor appear a line down. It no longer does that and can now be used for the text input cursor. It works really well.